set och get funkar men gameover highscore dåligt

+++ www/public/api/getHighscore.php
@@ -6,7 +6,7 @@
 
     // Kolla session med uid
     if(isset($_SESSION['uid'])){
-        $result = getHighscore($_SESSION['uid']);
+        $result = (int)getHighscore($_SESSION['uid']);
         //$result = (int)getHighscore('33f6649e-d72a-11ef-ba2d-0242ac141002');
     }else{
         $result = 0;
